Refactor code in src/Endpoints/RetailerAPI/Products.php, src/Requests/ListProductsRequest.php and tests/Feature/Endpoints/RetailerApi/ProductsTest.php
